Set up the front page
We can now add players to the game. When we click the 'Start Game'
button then we go to the page with the dice on it. From there we can see
all the players in the game.

Next step is to add the score and let the game end when we reach the
water mark.